+++ src/pins.h
@@ -5,11 +5,11 @@
 
 // Pins used
 LiquidCrystal lcd(7, 8, 9, 10, 11, 12);
-const int pin_led_green = 4;
-const int pin_led_blue = 5;
-const int pin_button_black = 6;
-const int pin_button_joystick = 13;
-const int pin_x_joystick = 1;
-const int pin_y_joystick = 0;
+const byte pin_led_green = 4;
+const byte pin_led_blue = 5;
+const byte pin_button_black = 6;
+const byte pin_button_joystick = 13;
+const byte pin_x_joystick = 1;
+const byte pin_y_joystick = 0;
 
 #endif // PINS_H
