+++ src/screens/screens.h
@@ -1,6 +1,8 @@
 #ifndef SCREENS_H
 #define SCREENS_H
 
+#include <Arduino.h>
+
 enum Screen
 {
     RandomizerInit = 0,
@@ -9,7 +11,7 @@ enum Screen
     FighterSelection = 3,
 };
 
-int screen_selected = Screen::RandomizerInit;
+byte screen_selected = Screen::RandomizerInit;
 bool screen_redraw = true;
 
 #endif // SCREENS_H
